refactor(oas31): disable try it out mechanism explicitly for webhooks
Refs #8474
