+++ src/core/plugins/oas31/spec-extensions/wrap-selectors.js
@@ -2,21 +2,7 @@
  * @prettier
  */
 
-/**
- * Selector wrapper maker the only wraps the passed selector
- * when spec is of OpenAPI 3.1.0 version.
- */
-const onlyOAS31Wrap =
-  (selector) =>
-  (oriSelector, system) =>
-  (state, ...args) => {
-    if (system.getSystem().specSelectors.isOAS31()) {
-      const result = selector(state, ...args)
-      return typeof result === "function" ? result(system) : result
-    } else {
-      return oriSelector(...args)
-    }
-  }
+import { createOnlyOAS31SelectorWrapper } from "../fn"
 
 export const isOAS3 =
   (oriSelector, system) =>
@@ -25,6 +11,8 @@ export const isOAS3 =
     return isOAS31 || oriSelector(...args)
   }
 
-export const selectLicenseUrl = onlyOAS31Wrap(() => (system) => {
-  return system.oas31Selectors.selectLicenseUrl()
-})
+export const selectLicenseUrl = createOnlyOAS31SelectorWrapper(
+  () => (system) => {
+    return system.oas31Selectors.selectLicenseUrl()
+  }
+)
